Add a params.pp file, to allow OS specific differences to be set here, and include it in init.pp archive defined type, and make sure with an order, that it is instantiated in the catalog before the defined type is executed.
The order of evaluation in the catalog is ensured by the
include line in init.pp, and the following two lines,
setting the order.

The defined variables specific for OpenBSD are:
$tarcmd string, defining a shell command to run tar
$digest_types hash, defining the supported hash types, and their
                    shell commands

The $tarcmd is used in extract.pp, the $digest_types is used
in download.pp

The only extra ::osfamily is OpenBSD, the rest is the values
taken in the default: case of the case statement.

The change in download.pp, uses the keys() function from stdlib.
Because of that, stdlib is added as dependency to metadata.json,
and .fixtures.yml. I chose an arbitrary version as minimal version
not too old, and not too recent. Otherwise, stdlib 3.X should
also work.

Added OpenBSD to the metadata.json, in order to get the specs
covering it too.